requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.2
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
frozenlist==1.5.0
html5lib==1.1
idna==3.10
isodate==0.7.2
lxml==5.3.0
multidict==6.1.0
oscrypto==1.3.0
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.1.0
python-bidi==0.6.3
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
sorl-thumbnail==12.11.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
xlwt==1.3.0
yarl==1.17.1
zeep==4.3.1
